Broke useSubmit to separate function
